Change of architecture, using sqlalchemy - orm
