+++ pywps/server/wsgi.py
@@ -1,2 +1,1 @@
-#possibility to import the WSGI application to ensure nice hiearchy for applications using PyWPS
-from pywps.server.app.main import ServerConnection
+from pywps.server.app import *
